Improve local testability of plugins in development (#12)
The newfound `--plugin-dir` provides an alternative mechanism for
testing and experimenting with plugins locally, while still in
development.

Though still undocumented, it clearly shows up in `Claude --help`. I've
incorporated it into the existing workflows because I find it more
convenient.

Moreover, I've removed some "obvious" instructions. Anthropic highly
recommends to respect Claude's inherent knowledge. The fact that plugins
are composed of commands, skills, etc. is trivially part of the context
of working with Claude plugins. So, it's a prime section to remove as it
contributed nothing profound or nontrivial.